[CoreSummit] [11226] working on Teams API
