+++ OpenStack Summit/CoreSummit/TeamMessageRequest.swift
@@ -1,0 +1,16 @@
+//
+//  TeamMessageRequest.swift
+//  OpenStack Summit
+//
+//  Created by Alsey Coleman Miller on 12/29/16.
+//  Copyright © 2016 OpenStack. All rights reserved.
+//
+
+import SwiftFoundation
+import AeroGearHttp
+import AeroGearOAuth2
+
+public extension Store {
+    
+    
+}
